+++ checkpoint/orbax/checkpoint/metadata/__init__.py
@@ -16,6 +16,10 @@
 
 # pylint: disable=g-importing-member, g-bad-import-order
 
+from orbax.checkpoint.metadata.checkpoint import CheckpointMetadata
+from orbax.checkpoint.metadata.checkpoint import CheckpointMetadataStore
+from orbax.checkpoint.metadata.checkpoint import checkpoint_metadata_store
+
 from orbax.checkpoint.metadata.sharding import ShardingMetadata
 from orbax.checkpoint.metadata.sharding import NamedShardingMetadata
 from orbax.checkpoint.metadata.sharding import SingleDeviceShardingMetadata